Fix error with init in peaks.js@2.0.0
