+++ package.json
@@ -60,7 +60,7 @@
     "konva": "^9.0.0",
     "lodash": "4.17.21",
     "manifesto.js": "^4.2.14",
-    "peaks.js": "^2.1.0",
+    "peaks.js": "2.0.0",
     "prop-types": "^15.6.2",
     "react-bootstrap": "1.6.4",
     "react-dnd": "^7.0.2",
