+++ src/containers/WaveformContainer.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { initializePeaks } from '../actions/peaks-instance';
+import { initManifest } from '../actions/manifest';
+import { retrieveStreamMedia } from '../actions/forms';
 import Waveform from '../components/Waveform';
 
 const WaveformContainer = (props) => {
@@ -9,32 +11,56 @@ const WaveformContainer = (props) => {
   const overView = React.createRef();
   const mediaPlayer = React.createRef();
 
-  const [manifestURL, setManifestURL] = React.useState(props.manifestURL);
-  const [canvasIndex, setCanvasIndex] = React.useState(props.canvasIndex);
+  const { streamMediaLoading } = useSelector((state) => state.forms.streamInfo);
+  const mediaInfo = useSelector((state) => state.manifest.mediaInfo);
+
+  const smData = useSelector((state) => state.structuralMetadata.smData);
 
   const dispatch = useDispatch();
+
+  React.useEffect(() => {
+    if(props.manifestURL) {
+      dispatch(initManifest(props.manifestURL, props.canvasIndex));
+    }
+  }, []);
+
+
+  React.useEffect(() => {
+    // When given a .m3u8 playlist, use HLS to stream media
+    if (mediaInfo.isStream) {
+      dispatch(
+        retrieveStreamMedia(mediaInfo.src, mediaPlayer.current, {
+          withCredentials: props.withCredentials,
+        })
+      );
+    }
+  }, [mediaInfo]);
 
   React.useEffect(() => {
     let peaksOptions = {
       keyboard: true,
       pointMarkerColor: '#006eb0',
       showPlayheadTime: true,
-      zoomWaveformColor: 'rgba(117, 117, 117, 1)',
-      overviewWaveformColor: 'rgba(117, 117, 117, 1)',
       timeLabelPrecision: 3,
-      containers: {
-        zoomview: zoomView.current,
-        overview: overView.current,
+      zoomview: {
+        container: zoomView.current,
+        waveformColor: 'rgba(117, 117, 117, 1)',
+      },
+      overview: {
+        container: overView.current,
+        waveformColor: 'rgba(117, 117, 117, 1)',
       },
       mediaElement: mediaPlayer.current,
       withCredentials: props.withCredentials,
+      player: null,
     };
-
-    dispatch(initializePeaks(
-      peaksOptions,
-      manifestURL,
-      canvasIndex,));
-  }, []);
+    if(!streamMediaLoading && smData != []) {
+      dispatch(initializePeaks(
+        peaksOptions,
+        smData,
+        props.canvasIndex,));
+    }
+  }, [streamMediaLoading, smData]);
 
   return (
     <section className="waveform-section" data-testid="waveform-container">
